% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_james.R
\docType{data}
\name{theme_james}
\alias{theme_james}
\title{Custom GGPlot2 Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_james
}
\description{
My custom GGPlot2 Theme
}
\keyword{datasets}
